Remove testWithParam and merge bug-14610 test files into single nsrt file
